requirements.txt
fastapi
uvicorn
httpx
google-genai
python-dotenv
truststore
statsbombpy
